+++ Form/Type/AbstractType.php
@@ -52,6 +52,8 @@ abstract class AbstractType extends SymfonyAbstractType implements TypeInterface
         if ($this->dataClass) {
             $defaults['data_class'] = $this->dataClass;
         }
+
+        $resolver->setDefaults($defaults);
     }
 
     /**
@@ -185,7 +187,7 @@ abstract class AbstractType extends SymfonyAbstractType implements TypeInterface
      */
     public function getNotifyDeactivateMails($formData = [])
     {
-        return (bool) $this->getNotifyMail($formData);
+        return !$this->getNotifyMail($formData);
     }
 
     /**
@@ -193,7 +195,7 @@ abstract class AbstractType extends SymfonyAbstractType implements TypeInterface
      */
     public function getCustomerDeactivateMails($formData = [])
     {
-        return (bool) $this->getCustomerMail($formData);
+        return !$this->getCustomerMail($formData);
     }
 
     /**