+++ Manager/FormManager.php
@@ -98,6 +98,7 @@ class FormManager
         $translation->setSendAttachments(self::getValue($data, 'sendAttachments'));
         $translation->setDeactivateNotifyMails(self::getValue($data, 'deactivateNotifyMails'));
         $translation->setDeactivateCustomerMails(self::getValue($data, 'deactivateCustomerMails'));
+        $translation->setReplyTo(self::getValue($data, 'replyTo'));
         $translation->setChanged(new \DateTime());
 
         // Add Translation to Form
@@ -142,7 +143,6 @@ class FormManager
             $fieldTranslation->setTitle(self::getValue($fieldData, 'title'));
             $fieldTranslation->setPlaceholder(self::getValue($fieldData, 'placeholder'));
             $fieldTranslation->setDefaultValue(self::getValue($fieldData, 'defaultValue'));
-
 
             // Field Options
             $prefix = 'options[';