+++ Provider/DynamicProvider.php
@@ -12,8 +12,16 @@ class DynamicProvider implements ListProviderInterface
      */
     public function getFieldDescriptors($webspace, $locale, $uuid)
     {
+        @trigger_error(
+            __method__ . '() use the new dynamic list provider.',
+            E_USER_DEPRECATED
+        );
+
         $fieldDescriptors = [
             'id' => $this->createFieldDescriptor('id', '', 'public.id'),
+            'uuid' => $this->createFieldDescriptor('uuid', '', 'uuid', true),
+            'webspaceKey' => $this->createFieldDescriptor('webspaceKey', '', 'webspaceKey', true),
+            'locale' => $this->createFieldDescriptor('locale', '', 'locale'),
             'firstName' => $this->createFieldDescriptor('firstName'),
             'lastName' => $this->createFieldDescriptor('lastName'),
             'email' => $this->createFieldDescriptor('email'),
@@ -33,8 +41,9 @@ class DynamicProvider implements ListProviderInterface
 
     /**
      * @param string $name
-     * @param bool $disabled
      * @param string $type
+     * @param string $translationKey
+     * @param bool $disabled
      *
      * @return DoctrineFieldDescriptor
      */