+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+ - BUGFIX     #80    Fixed getCustomerDeactivateMails and getNotifyDeactivateMails Method
+ - BUGFIX     #79    Fixed resolver of configureOptions in AbstractType
+
 ## 1.0.0-RC3
 
  - FEATURE    #74    Added mailchimp field type (requires https://github.com/drewm/mailchimp-api)